Version uptick and home page updates
