+++ src/app/constants/app-version.ts
@@ -1,1 +1,1 @@
-export const APP_VERSION = '3.27.3';
+export const APP_VERSION = '3.27.4';
