+++ src/app/pages/welcome/sections/intro.section.html
@@ -4,8 +4,8 @@
   Dive into the ultimate fan-made Fallout hub! Simulate iconic Pip-Boy
   interfaces right in your browser, build and share custom apps/games, or
   connect with real-world replicas for an immersive wasteland experience.
-  Created by fans for fans, collectors, cosplayers, and devs—join our growing
-  community and unlock endless Vault-Tec possibilities.
+  Created by fans for fans, collectors, cosplayers, and devs. Join our growing
+  community!
 </p>
 
 <div class="actions">
