+++ src/app/pages/welcome/sections/pip-boys/companions.section.html
