+++ src/app/pages/welcome/sections/pip-boys/companions.section.scss
@@ -1,4 +1,4 @@
-@use '../../../styles/variables' as var;
+@use '../../../../styles/variables' as var;
 
 :host {
   .pip-grid {
