+++ src/app/pages/welcome/sections/pip-boys/simulation.section.html
