+++ src/app/pages/welcome/sections/pip-boys/simulation.section.scss
@@ -1,5 +1,5 @@
-@use '../../../styles/variables' as var;
-@use '../../../styles/colors' as col;
+@use '../../../../styles/variables' as var;
+@use '../../../../styles/colors' as col;
 
 :host {
   .pip-grid {
